review and speedups for 'rtengine/PF_correct_RT.cc'

+++ rtengine/procparams.h
@@ -649,7 +649,7 @@ struct ColorAppearanceParams {
 struct DefringeParams {
     bool    enabled;
     double  radius;
-    float   threshold;
+    int     threshold;
     std::vector<double> huecurve;
 
     DefringeParams();
